Daily Build
Not too many things have been changed in the Github, mostly because I have been busy working in another project on a procedural terrain generation algorithm. I have succesfully produced Perlin noise in a 2D plane, yet exptending that to a third dimension in Unity and adjusting terrain colors as a result (lower elevations should be blue for water, and high ones should be white for snowy peaks) of thos elevation changes.

So as of now my successes are:
- Perlin noise 2D structure associated with grayscale color changes

What is looking promising:
- Mesh colliders for moving on terrain
- color changes based on heights

What is proving chalenging:
- Hieght change as a tresult of noise, and making that height change feel smooth and gradual over larger distances and scales.