Adding Terrain Generation to Project
Now that the Terrain Generation stuff feels stable in the other project I created I have pushed it onto the main project. The nature of the terrain can be adjusted with the parameters in the "MapGenerator" object and can be created using the "generate" button in that script.

Successes:
- Terrain Generation
- Color based on heights of the terrain
- Colliders for the character to interact within.

Short Term Goals:
- Scaling down the outside of the map to create the setting of an island
- Adjusting colors to feel more natural
- Increasing resolutions
- Roads on the map - potentially - remains to be seen how difficult that is.

Long Term Goals:
- Adding Wave Function collapse for targets for the character to engage with.
- Adding Ray tracing to identify when one of the security features can identify robbery Bob.
- Creating an "ID" system with scriptable objects for Robbery Bob to interact with in a standardized way.